Update content panel when change task date
